Put the low before the high.
It appears that when the OpenWeatherMap API returns its low temperature
and high temperature for the day, that the low means that it's the
lowest temperature expected in the 24 hour period that runs from
midnight to midnight. That means that typically the low for Monday is
the temperature on Monday at about 6 AM before the sun rises, and this
low temperature comes before the high, which happens in the afternoon
after the sun has been out all day. However, this display currently
displays the low after the high, meaning that in the typical case,
the low and high are not in chronological order.

By reversing the order, the high and low will more frequently be in
chronological order, paralleling the way that most other weather
forecasts report highs and lows.

Note: if tomorrow is forcasted to be much colder than today, and the
weather pattern is going to dramatically cool off in the afternoon or
overnight, then sometimes the low temperature returned by OpenWeatherMap
will actually be at midnight at the end of the day. Hence, this new
solution is only mostly chronological, but it should be chronological
more often that the current way.

This helps address lmarzen/esp32-weather-epd#99